# openapi/openapi.gen.yaml
# ⚠️ THIS FILE IS AUTOGENERATED. DO NOT EDIT MANUALLY.

openapi: 3.0.3
info:
  title: Perfect Pet Project API
  version: 0.1.0
  description: API для Perfect Pet Project
  contact:
    name: Elina Chinkina
    url: https://github.com/kudrmax
#servers:
#  - url: http://localhost:8080
#    description: Локальный сервер для разработки
paths:
  /api/1/posts/create_post:
    post:
      summary: Создать новый пост
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
            examples:
              good:
                summary: Good request body
                value:
                  text: Some text
              bad:
                summary: Bad request body
                value:
                  text: ''
      responses:
        '201':
          description: Пост создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/1/posts/feed:
    get:
      summary: Получить ленту постов
      operationId: getFeed
      responses:
        '200':
          description: Список постов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
    PostCreate:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
    Post:
      type: object
      required:
        - id
        - text
        - createdAt
      properties:
        id:
          type: integer
        text:
          type: string
          minLength: 1
        createdAt:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request (400)
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    InternalError:
      description: Internal Error (500)
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string